pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "procraft"
version = "0.0.1"
description = "ProCraft: chain-of-reasoning audio factorization (data pipeline + factorized codec)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "multiafx",
    "pretty_midi>=0.2.11",
    "pyFluidSynth>=1.3",
    "soundfile>=0.12",
    "mido>=1.3",
    "numpy>=1.24,<2.0",
    "scipy>=1.10",
    "pandas>=2.0",
    "tqdm>=4.65",
    "click>=8.1",
    "pyyaml>=6.0",
    "requests>=2.31",
]

[project.optional-dependencies]
llm = [
    "transformers>=4.45",
    "accelerate>=0.34",
    "huggingface_hub>=0.25",
]
serve = [
    "vllm>=0.8.5",
]
dev = ["pytest>=7.0", "pytest-cov", "ruff"]

[tool.setuptools.packages.find]
include = ["procraft_data*", "configs*"]
